packages/agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsletter-agent"
version = "0.1.0"
description = "LangGraph-based newsletter agent (gather → rank → write) for industry news."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.2",
  "langgraph>=0.2",
  "langchain-openai>=0.1",
  "pydantic>=2.0",
  "httpx>=0.27",
  "feedparser>=6.0",
  "jinja2>=3.1",
  "opentelemetry-api>=1.24",
  "markdown-it-py>=3.0",
  "bleach>=6.1",
  "python-dateutil>=2.8",
]

[project.scripts]
newsletter-agent = "newsletter_agent.cli:main"

[project.optional-dependencies]
azure_foundry = [
  "azure-identity>=1.17",
  "azure-ai-projects>=1.0.0b1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
